backend/requirements.txt
flask
flask-socketio
redis
happytransformer
python-dotenv
sendgrid
openai